Refactor Commands, parsers and UI based on person and project
