+++ src/main/java/loopin/projectbook/logic/parser/ProjectBookParser.java
@@ -8,25 +8,35 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import loopin.projectbook.commons.core.LogsCenter;
-import loopin.projectbook.logic.commands.AddOrgMemberCommand;
-import loopin.projectbook.logic.commands.AddProjectCommand;
-import loopin.projectbook.logic.commands.AddTeamMemberCommand;
-import loopin.projectbook.logic.commands.AddVolunteerCommand;
+import loopin.projectbook.logic.commands.personcommands.AddOrgMemberCommand;
+import loopin.projectbook.logic.commands.personcommands.AddProjectCommand;
+import loopin.projectbook.logic.commands.personcommands.AddTeamMemberCommand;
+import loopin.projectbook.logic.commands.personcommands.AddVolunteerCommand;
 import loopin.projectbook.logic.commands.ClearCommand;
 import loopin.projectbook.logic.commands.Command;
 import loopin.projectbook.logic.commands.DeleteCommand;
 import loopin.projectbook.logic.commands.EditCommand;
 import loopin.projectbook.logic.commands.ExitCommand;
-import loopin.projectbook.logic.commands.FindCommand;
+import loopin.projectbook.logic.commands.personcommands.FindCommand;
 import loopin.projectbook.logic.commands.HelpCommand;
 import loopin.projectbook.logic.commands.ListCommand;
-import loopin.projectbook.logic.commands.ProjectAssignCommand;
-import loopin.projectbook.logic.commands.ProjectListCommand;
-import loopin.projectbook.logic.commands.ProjectRemoveCommand;
-import loopin.projectbook.logic.commands.RemarkCommand;
-import loopin.projectbook.logic.commands.ResolveRemarkCommand;
-import loopin.projectbook.logic.commands.ViewProjectCommand;
+import loopin.projectbook.logic.commands.projectcommands.ProjectAssignCommand;
+import loopin.projectbook.logic.commands.projectcommands.ProjectListCommand;
+import loopin.projectbook.logic.commands.projectcommands.ProjectRemoveCommand;
+import loopin.projectbook.logic.commands.personcommands.RemarkCommand;
+import loopin.projectbook.logic.commands.personcommands.ResolveRemarkCommand;
+import loopin.projectbook.logic.commands.projectcommands.ViewProjectCommand;
 import loopin.projectbook.logic.parser.exceptions.ParseException;
+import loopin.projectbook.logic.parser.person.AddOrgMemberCommandParser;
+import loopin.projectbook.logic.parser.person.AddProjectCommandParser;
+import loopin.projectbook.logic.parser.person.AddTeamMemberCommandParser;
+import loopin.projectbook.logic.parser.person.AddVolunteerCommandParser;
+import loopin.projectbook.logic.parser.person.FindCommandParser;
+import loopin.projectbook.logic.parser.person.RemarkCommandParser;
+import loopin.projectbook.logic.parser.person.ResolveRemarkCommandParser;
+import loopin.projectbook.logic.parser.project.ProjectAssignCommandParser;
+import loopin.projectbook.logic.parser.project.ProjectRemoveCommandParser;
+import loopin.projectbook.logic.parser.project.ViewProjectCommandParser;
 
 /**
  * Parses user input.
