+++ src/main/java/loopin/projectbook/logic/parser/person/AddOrgMemberCommandParser.java
@@ -1,4 +1,4 @@
-package loopin.projectbook.logic.parser;
+package loopin.projectbook.logic.parser.person;
 
 import static loopin.projectbook.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static loopin.projectbook.logic.parser.CliSyntax.PREFIX_EMAIL;
@@ -15,7 +15,12 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Stream;
 
-import loopin.projectbook.logic.commands.AddOrgMemberCommand;
+import loopin.projectbook.logic.commands.personcommands.AddOrgMemberCommand;
+import loopin.projectbook.logic.parser.ArgumentMultimap;
+import loopin.projectbook.logic.parser.ArgumentTokenizer;
+import loopin.projectbook.logic.parser.Parser;
+import loopin.projectbook.logic.parser.ParserUtil;
+import loopin.projectbook.logic.parser.Prefix;
 import loopin.projectbook.logic.parser.exceptions.ParseException;
 import loopin.projectbook.model.person.Email;
 import loopin.projectbook.model.person.Name;
