+++ src/main/java/loopin/projectbook/logic/parser/person/ResolveRemarkCommandParser.java
@@ -1,4 +1,4 @@
-package loopin.projectbook.logic.parser;
+package loopin.projectbook.logic.parser.person;
 
 import static java.util.Objects.requireNonNull;
 import static loopin.projectbook.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
@@ -7,7 +7,11 @@ import java.util.Arrays;
 import java.util.List;
 
 import loopin.projectbook.commons.core.index.Index;
-import loopin.projectbook.logic.commands.ResolveRemarkCommand;
+import loopin.projectbook.logic.commands.personcommands.ResolveRemarkCommand;
+import loopin.projectbook.logic.parser.ArgumentMultimap;
+import loopin.projectbook.logic.parser.ArgumentTokenizer;
+import loopin.projectbook.logic.parser.Parser;
+import loopin.projectbook.logic.parser.ParserUtil;
 import loopin.projectbook.logic.parser.exceptions.ParseException;
 
 //TODO: FIX IMPLEMENTATION. It doesn't seem to resolve a remark completely.
