+++ src/main/java/loopin/projectbook/logic/parser/project/ProjectAssignCommandParser.java
@@ -1,31 +1,35 @@
-package loopin.projectbook.logic.parser;
+package loopin.projectbook.logic.parser.project;
 
 import static loopin.projectbook.logic.parser.CliSyntax.PREFIX_NAME;
 import static loopin.projectbook.logic.parser.CliSyntax.PREFIX_PROJECT;
 
 import loopin.projectbook.commons.core.index.Index;
-import loopin.projectbook.logic.commands.ProjectRemoveCommand;
+import loopin.projectbook.logic.commands.projectcommands.ProjectAssignCommand;
+import loopin.projectbook.logic.parser.ArgumentMultimap;
+import loopin.projectbook.logic.parser.ArgumentTokenizer;
+import loopin.projectbook.logic.parser.Parser;
+import loopin.projectbook.logic.parser.ParserUtil;
 import loopin.projectbook.logic.parser.exceptions.ParseException;
 import loopin.projectbook.model.project.ProjectName;
 
 /**
- * Parses input arguments and creates a new {@link ProjectRemoveCommand} object.
+ * Parses input arguments and creates a new {@link ProjectAssignCommand} object.
  *
  * Supported formats:
  * {@code INDEX project/PROJECT_NAME}
  * {@code n/NAME project/PROJECT_NAME}
  */
-public final class ProjectRemoveCommandParser implements Parser<ProjectRemoveCommand> {
+public final class ProjectAssignCommandParser implements Parser<ProjectAssignCommand> {
 
     @Override
-    public ProjectRemoveCommand parse(String args) throws ParseException {
+    public ProjectAssignCommand parse(String args) throws ParseException {
         final String normalized = " " + args.trim();
         ArgumentMultimap argMultimap = ArgumentTokenizer.tokenize(normalized, PREFIX_NAME, PREFIX_PROJECT);
 
         if (args.trim().startsWith(PREFIX_NAME.getPrefix())) {
             //name-based assignment
             String nameValue = argMultimap.getValue(PREFIX_NAME)
-                    .orElseThrow(() -> new ParseException(ProjectRemoveCommand.MESSAGE_USAGE));
+                    .orElseThrow(() -> new ParseException(ProjectAssignCommand.MESSAGE_USAGE));
             ProjectName projectName = argMultimap.getValue(PREFIX_PROJECT)
                     .map(n -> {
                         try {
@@ -34,9 +38,9 @@ public final class ProjectRemoveCommandParser implements Parser<ProjectRemoveCom
                             throw new RuntimeException(e);
                         }
                     })
-                    .orElseThrow(() -> new ParseException(ProjectRemoveCommand.MESSAGE_USAGE));
+                    .orElseThrow(() -> new ParseException(ProjectAssignCommand.MESSAGE_USAGE));
 
-            return new ProjectRemoveCommand(nameValue.trim(), projectName);
+            return new ProjectAssignCommand(nameValue.trim(), projectName);
         }
 
         String preamble = argMultimap.getPreamble().trim();
@@ -48,15 +52,15 @@ public final class ProjectRemoveCommandParser implements Parser<ProjectRemoveCom
                         throw new RuntimeException(e);
                     }
                 })
-                .orElseThrow(() -> new ParseException(ProjectRemoveCommand.MESSAGE_USAGE));
+                .orElseThrow(() -> new ParseException(ProjectAssignCommand.MESSAGE_USAGE));
 
         if (!preamble.isEmpty()) {
             // index-based assignment
             Index index = ParserUtil.parseIndex(preamble);
-            return new ProjectRemoveCommand(index, projectName);
+            return new ProjectAssignCommand(index, projectName);
         }
 
         Index index = ParserUtil.parseIndex(preamble);
-        return new ProjectRemoveCommand(index, projectName);
+        return new ProjectAssignCommand(index, projectName);
     }
 }
