+++ src/test/java/loopin/projectbook/logic/commands/AddCommandIntegrationTest.java
@@ -4,6 +4,8 @@ import static loopin.projectbook.logic.commands.CommandTestUtil.assertCommandFai
 import static loopin.projectbook.logic.commands.CommandTestUtil.assertCommandSuccess;
 import static loopin.projectbook.testutil.TypicalPersons.getTypicalProjectBook;
 
+import loopin.projectbook.logic.commands.personcommands.AddCommand;
+import loopin.projectbook.logic.commands.personcommands.AddVolunteerCommand;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
