+++ src/test/java/loopin/projectbook/logic/commands/ResolveRemarkCommandTest.java
@@ -6,6 +6,7 @@ import static loopin.projectbook.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
 import static loopin.projectbook.testutil.TypicalPersons.getTypicalProjectBook;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import loopin.projectbook.logic.commands.personcommands.ResolveRemarkCommand;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
