+++ src/test/java/loopin/projectbook/logic/parser/FindCommandParserTest.java
@@ -6,9 +6,10 @@ import static loopin.projectbook.logic.parser.CommandParserTestUtil.assertParseS
 
 import java.util.Arrays;
 
+import loopin.projectbook.logic.parser.person.FindCommandParser;
 import org.junit.jupiter.api.Test;
 
-import loopin.projectbook.logic.commands.FindCommand;
+import loopin.projectbook.logic.commands.personcommands.FindCommand;
 import loopin.projectbook.model.person.NameContainsKeywordsPredicate;
 
 public class FindCommandParserTest {
